Rename and restructure Settings folder
The Settings folder (within Assets) by default contains the
rendering settings for the Universap Render Pipeline (URP).
We rename it to "Config" to make it clearer that they are
technical configurations for the game, not gameplay settings
from the player's perspective.

We create a subfolder for "Rendering", so that other configs
can also be added later on to the same Config folder (e. g.
input layouts etc.).